Split actions and commands
Commands (with side effects) are moved to separate file.
Actions (object creators) are available in galleryActions.
#3
