+++ client/reducers/galleryReducers.js
