+++ client/views/PhotoToolbarView.js
@@ -3,7 +3,7 @@
  */
 import AmpersandView from 'ampersand-view';
 
-import {previousPhoto, nextPhoto, openThumbnails} from '../actions/galleryActions.js';
+import {previousPhoto, nextPhoto, openThumbnails} from '../commands/galleryCommands.js';
 import app from 'ampersand-app';
 import template from '../templates/views/photoToolbar.jade';
